NTP-2256 tak-export logging

+++ tak-integration/export/tak-export.sh
@@ -37,7 +37,11 @@ cd `dirname $0`
 # Check environment.
 #-----------------------------------------------------------
 mkdir -p ${tmpDir}
-rm ${logFile}
+
+if [ -s "$logFile" ]
+then
+   rm ${logFile}
+fi
 
 printlog "INFO" "Begin: ny tak-export" >> ${logFile}
 printlog "INFO" "Begin: environment check" >> ${logFile}
